Moved flr and faults into the recoverability folder
