+++ src/NServiceBus.Core/NServiceBus.Core.csproj
@@ -137,15 +137,15 @@
     <Compile Include="BusNotifications.cs" />
     <Compile Include="Extensibility\ExtendableOptions.cs" />
     <Compile Include="Extensibility\ExtendableOptionsExtensions.cs" />
-    <Compile Include="Faults\ConfigureError.cs" />
-    <Compile Include="Faults\FailedMessage.cs" />
-    <Compile Include="Faults\ErrorsNotifications.cs" />
-    <Compile Include="Faults\ErrorSubscribers.cs" />
+    <Compile Include="Recoverability\Faults\ConfigureError.cs" />
+    <Compile Include="Recoverability\Faults\FailedMessage.cs" />
+    <Compile Include="Recoverability\Faults\ErrorsNotifications.cs" />
+    <Compile Include="Recoverability\Faults\ErrorSubscribers.cs" />
     <Compile Include="Features\FeatureRunner.cs" />
     <Compile Include="Features\FeatureState.cs" />
-    <Compile Include="FirstLevelRetries\FirstLevelRetry.cs" />
-    <Compile Include="Faults\ErrorQueueSettings.cs" />
-    <Compile Include="Faults\StoreFaultsInErrorQueue.cs" />
+    <Compile Include="Recoverability\FirstLevelRetries\FirstLevelRetry.cs" />
+    <Compile Include="Recoverability\Faults\ErrorQueueSettings.cs" />
+    <Compile Include="Recoverability\Faults\StoreFaultsInErrorQueue.cs" />
     <Compile Include="Forwarding\ForwardingContext.cs" />
     <Compile Include="Forwarding\ForwardingToDispatchConnector.cs" />
     <Compile Include="Forwarding\InvokeForwardingPipelineBehavior.cs" />
@@ -248,8 +248,8 @@
     <Compile Include="Features\DisplayDiagnosticsForFeatures.cs" />
     <Compile Include="Features\ExposeSettings.cs" />
     <Compile Include="Features\PrerequisiteStatus.cs" />
-    <Compile Include="FirstLevelRetries\FirstLevelRetries.cs" />
-    <Compile Include="FirstLevelRetries\FirstLevelRetryPolicy.cs" />
+    <Compile Include="Recoverability\FirstLevelRetries\FirstLevelRetries.cs" />
+    <Compile Include="Recoverability\FirstLevelRetries\FirstLevelRetryPolicy.cs" />
     <Compile Include="Pipeline\BehaviorInstance.cs" />
     <Compile Include="Pipeline\IExecutor.cs" />
     <Compile Include="Pipeline\IndividualLimitThreadPoolExecutor.cs" />
@@ -355,7 +355,7 @@
     <Compile Include="Pipeline\StepEnded.cs" />
     <Compile Include="Scheduling\Scheduler.cs" />
     <Compile Include="Scheduling\TaskDefinition.cs" />
-    <Compile Include="Faults\MoveFaultsToErrorQueueBehavior.cs" />
+    <Compile Include="Recoverability\Faults\MoveFaultsToErrorQueueBehavior.cs" />
     <Compile Include="Recoverability\SecondLevelRetries\DefaultSecondLevelRetryPolicy.cs" />
     <Compile Include="Serialization\CustomSerialization.cs" />
     <Compile Include="Serialization\CustomSerializer.cs" />
@@ -378,7 +378,7 @@
     <Compile Include="Transports\Msmq\MsmqReceiveWithTransactionScopeBehavior.cs" />
     <Compile Include="Transports\ReceiveOptions.cs" />
     <Compile Include="Forwarding\ForwardReceivedMessages.cs" />
-    <Compile Include="FirstLevelRetries\FirstLevelRetriesBehavior.cs" />
+    <Compile Include="Recoverability\FirstLevelRetries\FirstLevelRetriesBehavior.cs" />
     <Compile Include="Pipeline\Contexts\HandlingContext.cs" />
     <Compile Include="Unicast\Behaviors\MessageProcessingAbortedException.cs" />
     <Compile Include="Unicast\Config\RegisterHandlersInOrder.cs" />
@@ -557,9 +557,8 @@
     <Compile Include="DependencyLifecycle.cs" />
     <Compile Include="Encryption\EncryptionMutator.cs" />
     <Compile Include="Encryption\IEncryptionService.cs" />
-    <Compile Include="Faults\FaultsHeaderKeys.cs" />
-    <Compile Include="Faults\FaultsQueueCreator.cs" />
-    <Compile Include="Faults\IManageMessageFailures.cs" />
+    <Compile Include="Recoverability\Faults\FaultsHeaderKeys.cs" />
+    <Compile Include="Recoverability\Faults\FaultsQueueCreator.cs" />
     <Compile Include="Hosting\Configuration\EndpointNameAttribute.cs" />
     <Compile Include="Performance\Statistics\SLA\EndpointSLAAttribute.cs" />
     <Compile Include="Hosting\Configuration\IConfigureThisEndpoint.cs" />
@@ -657,7 +656,7 @@
     <Compile Include="Unicast\Transport\ControlMessageFactory.cs" />
     <Compile Include="Unicast\Transport\FailedMessageProcessingEventArgs.cs" />
     <Compile Include="Unicast\Transport\Config\UseTransportExtensions.cs" />
-    <Compile Include="FirstLevelRetries\FlrStatusStorage.cs" />
+    <Compile Include="Recoverability\FirstLevelRetries\FlrStatusStorage.cs" />
     <Compile Include="Unicast\Transport\ThroughputLimiter.cs" />
     <Compile Include="Unicast\Transport\TransportReceiver.cs" />
     <Compile Include="Unicast\Transport\TransactionSettings.cs" />
