+++ src/NServiceBus.Core/Recoverability/Faults/ConfigureError.cs
