+++ src/NServiceBus.Core/Recoverability/Faults/ErrorsNotifications.cs
