+++ src/NServiceBus.Core/Recoverability/Faults/FailedMessage.cs
