+++ src/NServiceBus.Core/Recoverability/FirstLevelRetries/FirstLevelRetryPolicy.cs
@@ -1,4 +1,4 @@
-namespace NServiceBus.FirstLevelRetries
+namespace NServiceBus.Recoverability.FirstLevelRetries
 {
     class FirstLevelRetryPolicy
     {
