+++ src/NServiceBus.Core/Recoverability/FirstLevelRetries/FlrStatusStorage.cs
@@ -1,4 +1,4 @@
-﻿namespace NServiceBus.FirstLevelRetries
+﻿namespace NServiceBus.Recoverability.FirstLevelRetries
 {
     using System;
     using System.Collections.Concurrent;
